Parallax, OnFloorChange event and camera automation
Now the elevators don't use unityevent for new floor message, instead it's all connected to the action OnFloorChanged present in the GameManager.
Various changes done to camera system as well as parallax.

There are some issues regarding the reset of values due to the player restarting / dying that are not yet fixed